:memo: Update README, change LICENSE to CC BY-SA 4.0
This commit changes the README from the template repository's README to
a more specific README to the FOSSRIT Runbook. I also set up and added
badges for TravisCI and ReadTheDocs.org.

Additionally, I changed the license from the BSD 3-Clause License to a
Creative Commons license, which makes more sense since this repo mostly
deals with content, not code.

Signed-off-by: Justin W. Flory <git@jwf.io>