:memo: Add Ansible infrastructure docs (closes #6)
This commit adds docs to explain how the FOSSRIT/infrastructure
repository is set up and managed. This is high-level information about
the git repo and how it works. It doesn't include specific information
about how to run the playbooks since I think that makes more sense to
include in the FOSSRIT/infrastructure README or contributing guidelines.

Closes #6.

Signed-off-by: Justin W. Flory <git@jwf.io>