:package: Add requirements.txt for readthedocs.org
This commit adds a `requirements.txt` generated from the `Pipfile.lock`
file with the following command:

```sh
pipenv lock -r > requirements.txt
```

This is required so ReadTheDocs.org correctly imports our dependencies
when building our docs. Notably, this is for the
`sphinx-markdown-tables` library, which is not in the default
environment. Sad face.

Signed-off-by: Justin W. Flory <git@jwf.io>