:beetle: readthedocs: Add RTD config file, requirements file
ReadTheDocs.org builds are still failing. This follows the suggestion in
the build failure by using the `.readthedocs.yml` config file to specify
some of the build instructions.

Also, somehow the `sphinx-markdown-tables` dependency was omitted in the
`requirements.txt` file, which was the root problem I was battling.

Verification that this is _actually_ building now, srsly:

    https://readthedocs.org/projects/fossrit-runbook/builds/10419590/

Signed-off-by: Justin W. Flory <git@jwf.io>