:building_construction: website: Add instructions for dev environment
Contributes FOSSRIT/fossrit.github.io#86, but does not close it.

This commit migrates the existing content from the `CONTRIBUTING.md`
into the Runbook, so there is only one place to refer to for important
information about the website. Subsequent PRs will update the
fossrit.github.io repo to point to the Runbook once this commit is built
in the production docs site.

Note these docs cover a a modified workflow for the development
environment. This change is introduced in FOSSRIT/fossrit.github.io#88.

Closes FOSSRIT/fossrit.github.io#80.

^^I am closing that issue with this commit because it asks for a
non-trivial amount of documentation about how to set up the Jekyll stack
from scratch. This is something that varies on different developer
workstations and operating systems. If someone sets something up one
way, it may not be reproducible to how the production site is built.
Since we are using fully Free Software container tools to build the
development environment, I prefer to choose one and only one way to
standardize our development environment. If someone wishes to do it
another way, then that is their own prerogative.

Signed-off-by: Justin W. Flory <git@jwf.io>