:ocean: digitalocean: First draft (closes #19)
This commit adds a new docs page for our DigitalOcean infrastructure,
identifies financial responsibility, and what is currently running
there. This page is mostly screenshots, but I don't think there is much
to document here. I don't log into this interface that often and most of
what I look for is included in this doc.

Closes #19.

Signed-off-by: Justin W. Flory <git@jwf.io>