:space_invader: tech-team: Initial commit (closes #7)
This commit adds _two_ new docs, whooaaaa! This is a first step to
documenting what the FOSS@RIT Tech Team is, and also leaves a loose
blueprint of how to get involved with the Tech Team.

The first page (`community/tech-team`) focuses on defining what the
FOSS@RIT Tech Team is, what some of their responsibilities and projects
are, and how to get in touch with folks.

The second page (`policy/join-tech-team`) provides instructions both for
new volunteers and maintainers of the Tech Team on how to on-board new
people. It is a barebones page right now and I expect to add onto it
over time (like once the GitHub organization docs page is done). But
this puts down something for the time being, and it can be iterated on
over time.

Closes #7.

Signed-off-by: Justin W. Flory <git@jwf.io>