:heavy_check_mark: exhibition-checklist: Initial commit (closes #16)
This commit adds a checklist-style page for exhibitions, shows, and
conferences that FOSS@MAGIC attends. I tried to make this as
comprehensive as I could based on what I remember from how we organize
events. I also added some new requirements based on new functionality
with our website, e.g. add projects and events to the website in
advance.

This is a first start, and hopefully will be iterated on over time as we
attend more events.

Closes #16.

Signed-off-by: Justin W. Flory <git@jwf.io>