:ballot_box: community: Add Election Night Hackathon guide (closes #17)
This is an initial commit of an organizer's guide for the RIT Election
Night Hackathon, an annual tradition organized by FOSS@MAGIC each year
in November on U.S. election night.

I tried to give some context for the event and provide sequential steps
on things we do when we are pulling this together each year. Also, I
added a historical archive of past events to hopefully give some sort of
reference for future organizers about what we have done in the past.

Closes #17.

Signed-off-by: Justin W. Flory <git@jwf.io>