:octocat: github: Initial commit (closes #4, closes FOSSRIT/tasks#107) (#42)
This commit is a first-take at documenting the numerous intricacies of
how the FOSSRIT GitHub organization works. It is a massive brain dump of
the last 3-4 years of how I have maintained things. This is the first
time most of this information has ever been public because it was mostly
from a routine I created out of a habit that was constantly changing and
evolving.

_Anyways_, this is my attempt to slow the rate of change and create a
more standardized way of how we make changes to the FOSSRIT GitHub
organization. Ideally from this point on, changes to the FOSSRIT GitHub
organization will involve discussion and updating documentation.

Since this is a first take, I am curious for feedback on others on how
understandable this is, or if anything comes off as needlessly complex.

Closes #4.

ALSO RUNBOOK v1.0.0 IS OFFICIALLY DONE NOW YAY!!!!!!! :tada:

Closes FOSSRIT/tasks#107.

cc: @ritjoe @whenbellstoll @Nolski @Tjzabel

Signed-off-by: Justin W. Flory <git@jwf.io>