:detective: meta: Add CODEOWNERS rules (closes #3) (#48)
This commit adds a CODEOWNERS rule definition file. This automatically
requests specific groups of people to review PRs for certain changes. In
addition to automatically requesting review, it is also useful for
segmenting off different types of changes to the Runbook.

So, for tech-focused stuff, @FOSSRIT/tech-team gets tagged for review.
For more policy and governance-focused content, @FOSSRIT/foss-magic is
tagged for review.

Nifty!

Closes #3.

Signed-off-by: Justin W. Flory <git@jwf.io>