:memo: Add issue templates (closes #2)
This commit adds a set of issue templates based on what I anticipate
future people will open issues for on this repository. There are
templates for the following scenarios:

* Incorrect/outdated documentation
* Need for new documentation
* General improvement (e.g. for contributor experience or CI)
* Other (catch-all for anything else)

The issue templates help others structure their requests to the
FOSS@MAGIC team and also saves a little bit of time by using the set of
labels that I am in a habit of using for categorizing things.

Closes #2.

Signed-off-by: Justin W. Flory <git@jwf.io>